+++ open_web_calendar/convert/base.py
@@ -25,6 +25,7 @@ from open_web_calendar.calendars import (
 )
 from open_web_calendar.clean_html import clean_html
 from open_web_calendar.encryption import EmptyFernetStore, FernetStore
+from open_web_calendar.error import ResponseTooLarge
 
 
 def get_text_from_url(url):
@@ -35,7 +36,7 @@ def get_text_from_url(url):
 class ConversionStrategy:
     """Base class for conversions.
 
-    You can customize how calendars are retrived and if encryption is used.
+    You can customize how calendars are retrieved and if encryption is used.
     """
 
     # TODO: add as parameters
@@ -52,6 +53,7 @@ class ConversionStrategy:
         self.encryption = EmptyFernetStore() if encryption is None else encryption
         self.lock = RLock()
         self.components = []
+        self.seen_errors: set[tuple[str, str]] = set()
         self.get_text_from_url = get_text_from_url
         self.debug = debug
         self.created()
@@ -60,13 +62,19 @@ class ConversionStrategy:
         """Template method for subclasses."""
 
     def error(self, ty, err, tb, url):
-        tb_s = io.StringIO()
-        traceback.print_exception(ty, err, tb, file=tb_s)
-        return self.convert_error(
-            str(err) if self.debug else type(err).__name__,
-            url,
-            tb_s.getvalue() if self.debug else "",
-        )
+        traceback.print_exception(ty, err, tb)
+        error_str = str(err) if self.debug else type(err).__name__
+        key = (error_str, url)
+        with self.lock:
+            if key in self.seen_errors:
+                return None
+            self.seen_errors.add(key)
+            tb_s = ""
+            if self.debug:
+                sio = io.StringIO()
+                traceback.print_exception(ty, err, tb, file=sio)
+                tb_s = sio.getvalue()
+            return self.convert_error(error_str, url, tb_s)
 
     def convert_error(self, error: str, url: str, tb_s: str):
         """Tell the client more about the error."""
@@ -86,7 +94,7 @@ class ConversionStrategy:
                 pass  # no error should pass silently; import this
 
     def get_calendars_from_url(self, url: str) -> Calendars:
-        """Return a lis of calendars from a URL."""
+        """Return a list of calendars from a URL."""
         if self.encryption.is_encrypted(url):
             url = self.encryption.decrypt(url).url
             if url is None:
@@ -123,13 +131,18 @@ class ConversionStrategy:
             index, url = index_url
             calendars = self.get_calendars_from_url(url)
             self.collect_components_from(index, calendars)
+        except ResponseTooLarge:
+            # CLN-007 caps must surface as HTTP 413, not as an in-band error event.
+            raise
         except:
             ty, err, tb = sys.exc_info()
-            with self.lock:
-                self.components.append(self.error(ty, err, tb, url))
+            component = self.error(ty, err, tb, url)
+            if component is not None:
+                with self.lock:
+                    self.components.append(component)
 
     def collect_components_from(self, index: int, calendars: Calendars):
-        """Collect all the compenents from the calendar."""
+        """Collect all the components from the calendar."""
         raise NotImplementedError("to be implemented in subclasses")
 
     def merge(self):